fix the ... c++ complie error
